Andrews changes to City, Graph. 12/05

+++ City.h
@@ -9,17 +9,17 @@ using namespace std;
 class City {
 	public:
 		City();
-		City(string name, int latitude_, int longitude_);
+		City(string name, double latitude_, double longitude_);
 		string getName();
-        int getLatitude();
-        int getLongitude();
-        int get_distance(City target);
+        double getLatitude();
+        double getLongitude();
+        double get_distance(City target);
         void add_adj_city(City new_city);
 		// map<string, double> getAdj();
 	private:
 		string city_name;
-        int latitude;
-        int longitude;
+        double latitude;
+        double longitude;
         vector<City> adj_cities;
 		// map<string, double> adj_cities;
 };
