+++ data_processing.cpp
@@ -32,8 +32,8 @@ vector<vector<string>> data_processing::process_data(string filename) {
 
 City data_processing::create_city(vector<string> cells) {
     string city_name = cells[0];
-    int latitude = stoi(cells[6]);
-    int longitude = stoi(cells[7]);
+    double latitude = stod(cells[6]);
+    double longitude = stod(cells[7]);
     City new_city(city_name, latitude, longitude);
     return new_city;
 }
